update user and task by id models
