+++ project-2/src/model/user.js
@@ -33,37 +33,6 @@ class User{
 }
 
 
-class Todo{
-    constructor(id, user, todo) {
-      this.id = id;
-      this.user = user;
-      this.todo = todo;
-    }
-    //update task by id
-    static updateById(id, user, result){
-      db.query(
-        "UPDATE task SET id = ?, user = ?, todo = ? WHERE id = ?",
-        [todo.id, todo.user, id],
-        (err, res) => {
-          if (err) {
-            console.log("error: ", err);
-            result(null, err);
-            return;
-          }
-          if (res.affectedRows == 0) {
-            //not found
-            result({ kind: "not_found" }, null);
-            return;
-          }
-          console.log("task updated: ", { ...todo });
-          result(null, { ...todo });
-        }
-      );
-    }
-  
-  }
-
-module.exports = Todo;
 
 module.exports = User;
