+++ project-2/src/routes/user.js
@@ -1,0 +1,9 @@
+const express = require("express");
+const router = express.Router();
+const userController = require("../controllers/user");
+
+//update user by ID
+router.put('/:id', userController.update);
+
+
+module.exports = router;
